+++ notes/pages/CHANGELOG.md
@@ -13,6 +13,8 @@
 		- {{issue 1085}}
 	- BUGFIX: `clipping.is_inside_polygon()` function was incorrect
 		- {{issue 1094}}
+	- BUGFIX: `ClippingPolygon` raises exception clipping empty data
+		- {{issue 1096}}
 		-
 - ## Version 1.3.0 - 2024-05-01
   id:: 65e30c28-021e-4c24-ab6e-a9e9fa7c6a51